data_pipeline/requirements.txt
pybaseball==2.3.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3
python-dotenv==1.0.1
